refactor: IRoasRanking 타입을 provider.ts로 통합
overview.ts, platform.ts에 중복 선언되어 있던 IRoasRanking을
provider.ts 단일 출처로 이동하고 각 파일에서 re-export
